Fix y axis exception

+++ samples/funnyHats/js/index.js
@@ -7,6 +7,8 @@ let canvasContext = canvasOutput.getContext('2d');
 let streaming = false;
 let src = null;
 let cap = null;
+let hatDst = null;
+let maskDst = null;
 let faces = null;
 let classifier = null;
 const faceDetectionPath = 'haarcascade_frontalface_default.xml';
@@ -18,6 +20,8 @@ function startVideoProcessing() {
   src = new cv.Mat(video.height, video.width, cv.CV_8UC4);
   gray = new cv.Mat();
   cap = new cv.VideoCapture(video);
+  hatDst = new cv.Mat();
+  maskDst = new cv.Mat();
   faces = new cv.RectVector();
   classifier = new cv.CascadeClassifier();
 
@@ -41,8 +45,6 @@ function processVideo() {
     let begin = Date.now();
     // start processing
     cap.read(src);
-    hatSrc = hats[currentHat].src.clone();
-    mask = hats[currentHat].mask.clone();
     // detect faces
     cv.cvtColor(src, gray, cv.COLOR_RGBA2GRAY, 0);
     classifier.detectMultiScale(gray, faces,
@@ -51,30 +53,23 @@ function processVideo() {
       let face = faces.get(i);
       // draw hat
       let scaledWidth = parseInt(hats[currentHat].scale * face.width);
-      let scaledHeight = parseInt(hats[currentHat].scale * face.width);
+      let scaledHeight = parseInt(hats[currentHat].scale * face.height);
       let yOffset = Number(hats[currentHat].yOffset);
-      cv.resize(hatSrc, hatSrc, new cv.Size(scaledWidth, scaledHeight), 0, 0, cv.INTER_LINEAR);
-      cv.resize(mask, mask, new cv.Size(scaledWidth, scaledHeight), 0, 0, cv.INTER_LINEAR);
+      cv.resize(hatSrc, hatDst, new cv.Size(scaledWidth, scaledHeight), 0, 0, cv.INTER_LINEAR);
+      cv.resize(mask, maskDst, new cv.Size(scaledWidth, scaledHeight), 0, 0, cv.INTER_LINEAR);
       let y2 = face.y + Math.round(yOffset * face.height);
-      let y1 = y2 - hatSrc.rows;
-      let x1 = face.x + parseInt(face.width / 2 - hatSrc.cols / 2);
-      let x2 = x1 + hatSrc.cols;
-      console.log(face.width, face.height);
-      if (x2 < video.width && y1 >= 0) {
-        hatSrc.copyTo(src.rowRange(y1, y2).colRange(x1, x2), mask);
-      } else if (y1 < 0) {
-        let hatRoi = hatSrc.roi(new cv.Rect(0, -y1, scaledWidth, hatSrc.cols + y1));
-        let maskRoi = mask.roi(new cv.Rect(0, -y1, scaledWidth, mask.cols + y1));
+      let y1 = y2 - scaledHeight;
+      let x1 = face.x + parseInt(face.width / 2 - scaledWidth / 2);
+      let x2 = x1 + scaledWidth;
+      if (x2 < video.width && y1 > 0) {
+        hatDst.copyTo(src.rowRange(y1, y2).colRange(x1, x2), maskDst);
+      } else if (y1 < 0 && scaledHeight > -y1) {
+        let hatRoi = hatDst.roi(new cv.Rect(0, -y1, scaledWidth, scaledHeight + y1));
+        let maskRoi = maskDst.roi(new cv.Rect(0, -y1, scaledWidth, scaledHeight + y1));
         hatRoi.copyTo(src.rowRange(0, y2).colRange(x1, x2), maskRoi);
         hatRoi.delete(); maskRoi.delete();
       }
-      //hatSrc.copyTo(src.rowRange(0, hatSrc.rows).colRange(0, hatSrc.cols), mask);
-      let point1 = new cv.Point(face.x, face.y);
-      let point2 = new cv.Point(face.x + face.width, face.y + face.height);
-      cv.rectangle(src, point1, point2, [0, 255, 0, 255]);
     }
-    hatSrc.delete();
-    mask.delete();
     // draw output video
     cv.imshow('canvasOutput', src);
